+++ src/FileStream.h
@@ -14,15 +14,15 @@ typedef struct FileStreamContext {
     size_t bufferSize;
 } FileStreamContext;
 
-int file_stream_read(MemoryStream *stream, uint32_t offset, size_t size, void *outBuf);
-//int file_stream_write(MemoryStream *stream, uint32_t offset, size_t size, void *inBuf);
+int file_stream_read(MemoryStream *stream, uint64_t offset, size_t size, void *outBuf);
+//int file_stream_write(MemoryStream *stream, uint64_t offset, size_t size, void *inBuf);
 int file_stream_get_size(MemoryStream *stream, size_t *sizeOut);
 
 int file_stream_clone(MemoryStream *stream, MemoryStream *streamClone);
 int file_stream_trim(MemoryStream *stream, size_t trimAtStart, size_t trimAtEnd);
 //int file_stream_expand(MemoryStream *stream, size_t expandAtStart, size_t expandAtEnd);
 
-int file_stream_copy_data(MemoryStream *originStream, uint32_t originOffset, MemoryStream *targetStream, uint32_t targetOffset, size_t size);
+int file_stream_copy_data(MemoryStream *originStream, uint64_t originOffset, MemoryStream *targetStream, uint64_t targetOffset, size_t size);
 
 void file_stream_free(MemoryStream *stream);
